requirements.txt
flask==3.0.3
werkzeug==3.0.3
requests
ultralytics
openai
python-dotenv
